# Complete consultant registration profiles system with security and performance optimizations