… .agent-os/specs/2025-09-07-consultant-registration-profiles/spec-lite.md
… .agent-os/specs/2025-09-07-consultant-registration-profiles/spec.md
… .agent-os/specs/2025-09-07-consultant-registration-profiles/sub-specs/api-spec.md
… .agent-os/specs/2025-09-07-consultant-registration-profiles/sub-specs/database-schema.md
… .agent-os/specs/2025-09-07-consultant-registration-profiles/sub-specs/technical-spec.md
… .agent-os/specs/2025-09-07-consultant-registration-profiles/tasks.md
… consultant-app/index.html
… consultant-app/package.json
… consultant-app/public/vite.svg
… consultant-app/src/components/RelationshipList.tsx
… consultant-app/src/components/ui/dialog.tsx
… consultant-app/src/components/ui/input.tsx
… consultant-app/src/components/ui/radio-group.tsx
… consultant-app/src/components/ui/textarea.tsx
… consultant-app/src/hooks/use-toast.ts
… consultant-app/src/hooks/useAuth.ts
… consultant-app/src/hooks/useDashboard.ts
… consultant-app/src/index.css
… consultant-app/src/lib/api.ts
… consultant-app/src/lib/auth.ts
… consultant-app/src/lib/utils.ts
… consultant-app/src/main.tsx
… consultant-app/src/pages/Clients.tsx
… consultant-app/src/pages/Dashboard.tsx
… consultant-app/src/pages/Invitations.tsx
… consultant-app/src/pages/Landing.tsx
… consultant-app/src/pages/Login.tsx
… consultant-app/src/pages/Register.tsx
… consultant-app/tailwind.config.js
… consultant-app/tsconfig.json